requirements.txt
groq>=0.11.0
google-generativeai>=0.8.0
openai>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
feedparser>=6.0.11
PyYAML>=6.0.1
beautifulsoup4>=4.12.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pypdf>=5.0.0
